Refactor ProgramOptions struct
Refactor it to include the configuration options for both generate_table
and crack_hash.

Originally it was intended for the operation_options field to be of a
different type, depending on the operation. However it does not seem
easy to do so the ProgramOptions struct now contains all configuration
options for both operations, however these will be optional. Depending
on the desired operation, the program will parse and validate the
arguments and fill in the optional values